Various cleanups to avoid compiler warnings

+++ SAX2.c
@@ -1889,7 +1889,7 @@ skip:
 	    else {
 	        ret->line = 65535;
 		if (ctxt->options & XML_PARSE_BIG_LINES)
-		    ret->psvi = (void *) ctxt->input->line;
+		    ret->psvi = (void *) (long) ctxt->input->line;
 	    }
 	}
     }
